machine_learning/requirements.txt
Flask
Pillow
requests
tensorflow